almost finish introduction fragment

+++ base/src/main/java/com/fengsheng/base/network/bean/PopularVideoDataBean.kt
@@ -58,7 +58,7 @@ data class PopularVideoDataBean(val code: Int, val data: Data, val message: Stri
             data class Stat(
                 val aid: Int,
                 val coin: Int,
-                val danmaku: Int,
+                val danmaku: Long,
                 val dislike: Int,
                 val favorite: Int,
                 val his_rank: Int,
